refact(community): change in screen file
